Sammen spil med de 3 sider
create-store-comp, create-product comp og create-prouctlist-comp
