+++ spar-med-samvittighed/src/app/store/[storeId]/page.tsx
@@ -7,11 +7,13 @@ interface StoreProp {
   };
 }
 
-export default function Products({ params }: StoreProp) {
+const Products: React.FC<StoreProp> = ({ params }) => {
   return (
     <div className="flex min-h-screen flex-col items-center justify-between p-24">
-        <h1>Products from storeid {params.storeId}</h1>
-        <StoreProductList />
+      <h1>Products from store ID: {params.storeId}</h1>
+      <StoreProductList storeId={params.storeId} />
     </div>
   );
-}
+};
+
+export default Products;
